requirements.txt
fastapi
uvicorn
librosa
numpy
python-dotenv
openai-whisper
torch
ffmpeg-python
openai
